+++ app/Http/Controllers/ResourceController.php
@@ -11,6 +11,8 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
 use App\Http\Controllers\Api\ResourceController as ApiResource;
 use App\Http\Controllers\Api\ConversionController as ApiConversion;
+use App\Http\Controllers\Api\ResourceController as ApiResource;
+use App\Http\Controllers\Api\ConversionController as ApiConversion;
 
 class ResourceController extends Controller {
     public static function addMetadata($recordUri, $resourceData, $file = null, $isUpdate = false)
@@ -577,4 +579,33 @@ class ResourceController extends Controller {
         return back();
     }
 
+    public function execResourceQueryScript(Request $request)
+    {
+        $format = $request->format;
+        $resourceParams = ['resource_uri' => $request->uri, 'version' => $request->version];
+
+        $rq = Request::create('/api/getResourceData', 'POST', $resourceParams);
+        $api = new ApiResource($rq);
+        $res = $api->getResourceData($rq)->getData();
+
+        if ($res->success) {
+            $data = isset($res->data) ? $res->data : [];
+
+            if ($format == Page::RESOURCE_RESPONSE_CSV) {
+                $convertData = ['data' => $data];
+                $reqConvert = Request::create('/json2csv', 'POST', $convertData);
+                $apiConvert = new ApiConversion($reqConvert);
+                $resultConvert = $apiConvert->json2csv($reqConvert)->getData();
+                $data = isset($resultConvert->data) ? $resultConvert->data : [];
+                $res->success = $resultConvert->success;
+            }
+        } else {
+            $data = isset($res->errors) ? $res->errors : [];
+        }
+
+        return [
+            'success' => $res->success,
+            'data'    => json_encode($data)
+        ];
+    }
 }